Create GRID3.html
This is the third attempt at creating the grid. It is now transposed with levels increasing from left to right. The different sections of the curriculum are from top to bottom. This allows us to colour the rows in relation to a strand and then colour grids in relation to the substrands.