Update GRID3.html
This is all of the buttons in place. The colours are still ugly and there is a whole lot of other CSS to do. I also need to remove that gap between level 6 and 7. That is an easy fix to move the title of the substrand to the right for those that start after level 6. 

I have investigated hover. It can be big enough to have the content and ordered list of elaborations. 

It will best if these are variables that we call on with DOM. Same with the learning task list. 

I am still a long way away from knowing how to mange the data, let alone turn it into some sort of visualisation.